Rename dmscripts module
