+++ dmscripts/export_framework_applications_at_close.py
